Introduce time-driven spiking connections
The framework for continous interactions is abused to communicate
spikes. The corresponding synapses are updated in every time
step. While this is much less efficient than the traditional
event-driven approach used in NEST, it supports the fast
implementation of various plasticity rules in NEST that rely on
continously updating synaptic state variables. As a proof of concept
the iaf_psc_exp neuron now can be connected via a new
time_diven_static_connection, which results in identical dynamics as
the matching event-driven implementation.